+++ src/GRA.Controllers/ViewModel/MissionControl/Participants/PrizeListViewModel.cs
@@ -5,8 +5,12 @@ namespace GRA.Controllers.ViewModel.MissionControl.Participants
 {
     public class PrizeListViewModel : ParticipantPartialViewModel
     {
-        public IEnumerable<GRA.Domain.Model.PrizeWinner> PrizeWinners { get; set; }
+        public IEnumerable<Domain.Model.PrizeWinner> PrizeWinners { get; set; }
         public PaginateViewModel PaginateModel { get; set; }
         public int? HeadOfHouseholdId { get; set; }
+        public Domain.Model.User User { get; set; }
+        public bool CanEditDetails { get; set; }
+
+        public EmailAwardViewModel EmailAwardModel { get; set; }
     }
 }