+++ src/GRA.Domain.Model/Filters/EmailReminderFilter.cs
@@ -1,0 +1,8 @@
+﻿namespace GRA.Domain.Model.Filters
+{
+    public class EmailReminderFilter : BaseFilter
+    {
+        public string MailingList { get; set; }
+        public bool HasReceived { get; set; }
+    }
+}