+++ src/GRA.Domain.Service/UserService.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Diagnostics;
+using System.Globalization;
 using System.IO;
 using System.Linq;
 using System.Text;
@@ -241,7 +242,8 @@ namespace GRA.Domain.Service
             {
                 if (!string.IsNullOrEmpty(filter.Search))
                 {
-                    var vendorCode = await _vendorCodeService.GetVendorCodeByCode(filter.Search.ToUpper());
+                    var vendorCode = await _vendorCodeService
+                        .GetVendorCodeByCode(filter.Search.ToUpper(CultureInfo.InvariantCulture));
                     if (vendorCode?.UserId.HasValue == true)
                     {
                         filter.UserIds = new List<int>
@@ -324,23 +326,34 @@ namespace GRA.Domain.Service
             int authUserId = GetClaimId(ClaimType.UserId);
             var authUser = await _userRepository.GetByIdAsync(authUserId);
             var requestedUser = await _userRepository.GetByIdAsync(userId);
-            if (requestedUser == null)
-            {
-                throw new GraException("The requested participant could not be accessed or does not exist.");
-            }
-            if (authUserId == userId
-                || requestedUser.HouseholdHeadUserId == authUserId
-                || authUser.HouseholdHeadUserId == userId
-                || HasPermission(Permission.ViewParticipantDetails))
+
+            if (requestedUser != null
+                && (authUserId == userId
+                    || requestedUser.HouseholdHeadUserId == authUserId
+                    || authUser.HouseholdHeadUserId == userId))
             {
                 return requestedUser;
             }
             else
             {
-                _logger.LogError("User {UserId} doesn't have permission to view participant details",
-                    authUserId);
+                _logger.LogError("User {AuthUserId} is not allowed to view participant details of {UserId}",
+                    authUserId,
+                    userId);
                 throw new GraException(_sharedLocalizer[Annotations.Validate.Permission]);
             }
+        }
+
+        public async Task<User> GetDetailsByPermission(int userId)
+        {
+            VerifyPermission(Permission.ViewParticipantDetails);
+
+            var requestedUser = await _userRepository.GetByIdAsync(userId);
+            if (requestedUser == null)
+            {
+                throw new GraException("The requested participant could not be accessed or does not exist.");
+            }
+
+            return requestedUser;
         }
 
         public async Task<User> Update(User userToUpdate)
@@ -597,7 +610,7 @@ namespace GRA.Domain.Service
         public async Task<string>
             ActivateAuthorizationCode(string authorizationCode, int? joiningUserId = null)
         {
-            string fixedCode = authorizationCode.Trim().ToLower();
+            string fixedCode = authorizationCode.Trim().ToLower(CultureInfo.InvariantCulture);
             int siteId = GetCurrentSiteId();
             var authCode
                 = await _authorizationCodeRepository.GetByCodeAsync(siteId, fixedCode);
@@ -643,8 +656,8 @@ namespace GRA.Domain.Service
                 var rolePermissions = (await _roleRepository
                     .GetPermissionNamesForRoleAsync(authCode.RoleId)).ToList();
 
-                if (rolePermissions.Contains(Permission.NewsAutoSubscribe.ToString())
-                    && (rolePermissions.Contains(Permission.AccessMissionControl.ToString())
+                if (rolePermissions.Contains(nameof(Permission.NewsAutoSubscribe))
+                    && (rolePermissions.Contains(nameof(Permission.AccessMissionControl))
                         || HasPermission(Permission.AccessMissionControl)))
                 {
                     user.IsNewsSubscribed = true;
@@ -1478,11 +1491,6 @@ namespace GRA.Domain.Service
             }
         }
 
-        public async Task<int> GetSystemUserId()
-        {
-            return await _userRepository.GetSystemUserId();
-        }
-
         public async Task<DataWithCount<ICollection<GroupInfo>>> GetPaginatedGroupListAsync(
             GroupFilter filter)
         {
@@ -1580,7 +1588,6 @@ namespace GRA.Domain.Service
             {
                 var sw = new Stopwatch();
                 sw.Start();
-
 
                 var job = await _jobRepository.GetByIdAsync(jobId);
                 var jobDetails
@@ -1761,7 +1768,6 @@ namespace GRA.Domain.Service
                     Status = "<strong>Import Complete</strong>"
                 };
             }
-
             else
             {
                 _logger.LogError("User {UserId} doesn't have permission to import household members.",
@@ -1774,6 +1780,11 @@ namespace GRA.Domain.Service
                     Complete = true
                 };
             }
+        }
+
+        public async Task<bool> IsEmailSubscribedAsync(string email)
+        {
+            return await _userRepository.IsEmailSubscribedAsync(email);
         }
     }
 }